Function Update
I have created functions which split the file up with the ability to add in additional functions. They generally wrok well but there are some bugs where the headers are not insterted, I need to find out why.
I also updated filtertime to make it simpler removong the breaking up as that wasn't working well.