ConvertToTime
I made a working function that can convert different time series into Posix when they are all in the same column. This still needs to be put into it's own file

Also changed the RDS load/save to CSV and converted the parse of a text function into the evaluation of a subsititute function.